# crow_eye_api/routers/audio.py
"""
Audio router for Crow's Eye API.
"""

import os
import sys
import shutil
from pathlib import Path
from typing import List, Optional
from fastapi import APIRouter, UploadFile, File, HTTPException, Depends, status
from pydantic import BaseModel

# Add src to path for imports
sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..', '..', 'src'))

from src.models.user import User
from src.features.subscription.access_control import Feature
from ..dependencies import get_current_user, require_feature

router = APIRouter()

# Audio storage directory
AUDIO_DIR = Path("data/audio")
AUDIO_DIR.mkdir(parents=True, exist_ok=True)


class AudioImportRequest(BaseModel):
    """Audio import request model."""
    name: Optional[str] = None
    description: Optional[str] = None
    tags: Optional[List[str]] = None
    auto_enhance: Optional[bool] = True
    normalize_volume: Optional[bool] = True


class AudioResponse(BaseModel):
    """Audio response model."""
    id: str
    filename: str
    name: str
    description: Optional[str]
    duration: float  # seconds
    format: str
    size: int  # bytes
    sample_rate: int
    channels: int
    tags: List[str]
    created_at: str
    url: str
    waveform_url: Optional[str]


class AudioListResponse(BaseModel):
    """Audio list response model."""
    audio_files: List[AudioResponse]
    total: int


class AudioEditRequest(BaseModel):
    """Audio editing request model."""
    audio_id: str
    operations: List[dict]  # List of editing operations
    output_format: Optional[str] = "mp3"


@router.post("/import", response_model=AudioResponse)
async def import_audio(
    file: UploadFile = File(...),
    request: AudioImportRequest = AudioImportRequest(),
    current_user: User = Depends(require_feature(Feature.AUDIO_IMPORTER))
):
    """
    Import custom audio file with natural language editing capabilities.
    
    **Tier Required:** Creator+ ($9/month and above)
    
    Supports audio formats:
    - MP3, WAV, FLAC, AAC, OGG
    - Automatic format conversion
    - Volume normalization
    - Audio enhancement
    """
    # Validate file type
    allowed_types = {
        "audio/mpeg", "audio/wav", "audio/flac", 
        "audio/aac", "audio/ogg", "audio/x-wav"
    }
    
    if file.content_type not in allowed_types:
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail=f"Unsupported audio format: {file.content_type}"
        )
    
    # Generate unique filename
    import uuid
    from datetime import datetime
    
    file_extension = Path(file.filename).suffix
    unique_filename = f"{uuid.uuid4()}{file_extension}"
    file_path = AUDIO_DIR / unique_filename
    
    # Save file
    try:
        with open(file_path, "wb") as buffer:
            shutil.copyfileobj(file.file, buffer)
    except Exception as e:
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail=f"Failed to save audio file: {str(e)}"
        )
    
    # Mock audio analysis
    # In production, you'd use libraries like librosa or pydub to analyze the audio
    file_stats = file_path.stat()
    
    return AudioResponse(
        id=str(uuid.uuid4()),
        filename=file.filename,
        name=request.name or Path(file.filename).stem,
        description=request.description,
        duration=120.5,  # Mock duration
        format=file_extension.lstrip('.').upper(),
        size=file_stats.st_size,
        sample_rate=44100,
        channels=2,
        tags=request.tags or [],
        created_at=datetime.now().isoformat(),
        url=f"/audio/{unique_filename}",
        waveform_url=f"/audio/{unique_filename}/waveform"
    )


@router.get("/{audio_id}", response_model=AudioResponse)
async def get_audio(
    audio_id: str,
    current_user: User = Depends(require_feature(Feature.AUDIO_IMPORTER))
):
    """
    Get audio file information by ID.
    
    **Tier Required:** Creator+ ($9/month and above)
    """
    # This is a simplified implementation
    # In production, you'd look up the audio by ID in a database
    
    from datetime import datetime
    
    return AudioResponse(
        id=audio_id,
        filename="sample_audio.mp3",
        name="Sample Audio",
        description="A sample audio file",
        duration=180.0,
        format="MP3",
        size=2048000,
        sample_rate=44100,
        channels=2,
        tags=["music", "background"],
        created_at=datetime.now().isoformat(),
        url=f"/audio/{audio_id}",
        waveform_url=f"/audio/{audio_id}/waveform"
    )


@router.get("/", response_model=AudioListResponse)
async def list_audio(
    limit: int = 50,
    offset: int = 0,
    tags: Optional[str] = None,  # comma-separated tags
    current_user: User = Depends(require_feature(Feature.AUDIO_IMPORTER))
):
    """
    List user's audio files.
    
    **Tier Required:** Creator+ ($9/month and above)
    """
    # This is a simplified implementation
    # In production, you'd query the database for user's audio files
    
    from datetime import datetime
    
    # Mock audio list
    audio_files = []
    for i in range(min(limit, 10)):  # Return up to 10 mock items
        audio_files.append(AudioResponse(
            id=f"audio_{i + offset}",
            filename=f"audio_{i + offset}.mp3",
            name=f"Audio Track {i + offset + 1}",
            description=f"Sample audio track {i + offset + 1}",
            duration=120.0 + i * 30,
            format="MP3",
            size=2048000 + i * 100000,
            sample_rate=44100,
            channels=2,
            tags=["music", "background"] if i % 2 == 0 else ["voice", "narration"],
            created_at=datetime.now().isoformat(),
            url=f"/audio/audio_{i + offset}.mp3",
            waveform_url=f"/audio/audio_{i + offset}/waveform"
        ))
    
    return AudioListResponse(
        audio_files=audio_files,
        total=len(audio_files)
    )


@router.post("/edit", response_model=AudioResponse)
async def edit_audio(
    request: AudioEditRequest,
    current_user: User = Depends(require_feature(Feature.AUDIO_IMPORTER))
):
    """
    Edit audio with natural language commands.
    
    **Tier Required:** Creator+ ($9/month and above)
    
    Supported operations:
    - Trim: "Cut from 0:30 to 1:45"
    - Volume: "Increase volume by 20%"
    - Fade: "Add fade in for 3 seconds"
    - Normalize: "Normalize audio levels"
    - Effects: "Add reverb", "Remove background noise"
    """
    # This is a simplified implementation
    # In production, you'd process the audio using libraries like pydub or ffmpeg
    
    import uuid
    from datetime import datetime
    
    # Mock audio editing
    # Parse natural language operations and apply them
    
    edited_audio_id = str(uuid.uuid4())
    
    return AudioResponse(
        id=edited_audio_id,
        filename=f"edited_{request.audio_id}.{request.output_format}",
        name=f"Edited Audio {request.audio_id}",
        description="Edited audio file",
        duration=150.0,  # Mock edited duration
        format=request.output_format.upper(),
        size=1800000,
        sample_rate=44100,
        channels=2,
        tags=["edited", "processed"],
        created_at=datetime.now().isoformat(),
        url=f"/audio/{edited_audio_id}",
        waveform_url=f"/audio/{edited_audio_id}/waveform"
    )


@router.delete("/{audio_id}")
async def delete_audio(
    audio_id: str,
    current_user: User = Depends(require_feature(Feature.AUDIO_IMPORTER))
):
    """
    Delete audio file by ID.
    
    **Tier Required:** Creator+ ($9/month and above)
    """
    # This is a simplified implementation
    # In production, you'd delete from database and file system
    
    return {"message": f"Audio {audio_id} deleted successfully"}


class AudioEffectResponse(BaseModel):
    """Audio effect response model."""
    id: str
    name: str
    description: str
    parameters: List[dict]
    category: str


@router.get("/effects/", response_model=List[AudioEffectResponse])
async def list_audio_effects(
    current_user: User = Depends(require_feature(Feature.AUDIO_IMPORTER))
):
    """
    List available audio effects.
    
    **Tier Required:** Creator+ ($9/month and above)
    """
    # Mock effects list
    effects = [
        AudioEffectResponse(
            id="reverb",
            name="Reverb",
            description="Add spatial depth and ambiance",
            parameters=[
                {"name": "room_size", "type": "float", "min": 0.0, "max": 1.0, "default": 0.5},
                {"name": "damping", "type": "float", "min": 0.0, "max": 1.0, "default": 0.3}
            ],
            category="spatial"
        ),
        AudioEffectResponse(
            id="noise_reduction",
            name="Noise Reduction",
            description="Remove background noise and hiss",
            parameters=[
                {"name": "strength", "type": "float", "min": 0.0, "max": 1.0, "default": 0.7}
            ],
            category="cleanup"
        ),
        AudioEffectResponse(
            id="equalizer",
            name="Equalizer",
            description="Adjust frequency response",
            parameters=[
                {"name": "bass", "type": "float", "min": -20.0, "max": 20.0, "default": 0.0},
                {"name": "mid", "type": "float", "min": -20.0, "max": 20.0, "default": 0.0},
                {"name": "treble", "type": "float", "min": -20.0, "max": 20.0, "default": 0.0}
            ],
            category="tone"
        ),
        AudioEffectResponse(
            id="compressor",
            name="Compressor",
            description="Control dynamic range",
            parameters=[
                {"name": "threshold", "type": "float", "min": -60.0, "max": 0.0, "default": -20.0},
                {"name": "ratio", "type": "float", "min": 1.0, "max": 20.0, "default": 4.0}
            ],
            category="dynamics"
        )
    ]
    
    return effects


@router.get("/{audio_id}/analyze")
async def analyze_audio(
    audio_id: str,
    current_user: User = Depends(require_feature(Feature.AUDIO_IMPORTER))
):
    """
    Analyze audio file properties and characteristics.
    
    **Tier Required:** Creator+ ($9/month and above)
    """
    # Mock audio analysis
    return {
        "id": audio_id,
        "analysis": {
            "peak_volume": -3.2,
            "rms_volume": -18.5,
            "dynamic_range": 15.3,
            "frequency_spectrum": {
                "bass": 0.3,
                "mid": 0.6,
                "treble": 0.4
            },
            "tempo": 120,  # BPM
            "key": "C major",
            "mood": "upbeat",
            "genre": "electronic",
            "quality_score": 8.5
        },
        "recommendations": [
            "Consider normalizing volume levels",
            "Audio quality is excellent",
            "Good dynamic range for music"
        ]
    } 