# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Crow's Eye Marketing Platform API",
    "description": "REST API for the Crow's Eye marketing automation platform",
    "version": "1.0.0"
  },
  "paths": {
...
    "/audio/import": {
      "post": {
        "tags": [
          "Audio"
        ],
        "summary": "Import Audio",
        "description": "Import custom audio file with natural language editing capabilities.\n\n**Tier Required:** Creator+ ($9/month and above)\n\nSupports audio formats:\n- MP3, WAV, FLAC, AAC, OGG\n- Automatic format conversion\n- Volume normalization\n- Audio enhancement",
        "operationId": "import_audio_audio_import_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-user-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_audio_audio_import_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/audio/{audio_id}": {
      "get": {
        "tags": [
          "Audio"
        ],
        "summary": "Get Audio",
        "description": "Get audio file information by ID.\n\n**Tier Required:** Creator+ ($9/month and above)",
        "operationId": "get_audio_audio__audio_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "audio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Audio Id"
            }
          },
          {
            "name": "x-user-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Audio"
        ],
        "summary": "Delete Audio",
        "description": "Delete audio file by ID.\n\n**Tier Required:** Creator+ ($9/month and above)",
        "operationId": "delete_audio_audio__audio_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "audio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Audio Id"
            }
          },
          {
            "name": "x-user-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/audio/": {
      "get": {
        "tags": [
          "Audio"
        ],
        "summary": "List Audio",
        "description": "List user's audio files.\n\n**Tier Required:** Creator+ ($9/month and above)",
        "operationId": "list_audio_audio__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "tags",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tags"
            }
          },
          {
            "name": "x-user-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/audio/edit": {
      "post": {
        "tags": [
          "Audio"
        ],
        "summary": "Edit Audio",
        "description": "Edit audio with natural language commands.\n\n**Tier Required:** Creator+ ($9/month and above)\n\nSupported operations:\n- Trim: \"Cut from 0:30 to 1:45\"\n- Volume: \"Increase volume by 20%\"\n- Fade: \"Add fade in for 3 seconds\"\n- Normalize: \"Normalize audio levels\"\n- Effects: \"Add reverb\", \"Remove background noise\"",
        "operationId": "edit_audio_audio_edit_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-user-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AudioEditRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/audio/effects/": {
      "get": {
        "tags": [
          "Audio"
        ],
        "summary": "List Audio Effects",
        "description": "List available audio effects.\n\n**Tier Required:** Creator+ ($9/month and above)",
        "operationId": "list_audio_effects_audio_effects__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "x-user-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AudioEffectResponse"
                  },
                  "title": "Response List Audio Effects Audio Effects  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/audio/{audio_id}/analyze": {
      "get": {
        "tags": [
          "Audio"
        ],
        "summary": "Analyze Audio",
        "description": "Analyze audio file properties and characteristics.\n\n**Tier Required:** Creator+ ($9/month and above)",
        "operationId": "analyze_audio_audio__audio_id__analyze_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "audio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Audio Id"
            }
          },
          {
            "name": "x-user-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AudioEditRequest": {
        "properties": {
          "audio_id": {
            "type": "string",
            "title": "Audio Id"
          },
          "operations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Operations"
          },
          "output_format": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Format",
            "default": "mp3"
          }
        },
        "type": "object",
        "required": [
          "audio_id",
          "operations"
        ],
        "title": "AudioEditRequest",
        "description": "Audio editing request model."
      },
      "AudioEffectResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "parameters": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Parameters"
          },
          "category": {
            "type": "string",
            "title": "Category"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "parameters",
          "category"
        ],
        "title": "AudioEffectResponse",
        "description": "Audio effect response model."
      },
      "AudioImportRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "auto_enhance": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Enhance",
            "default": true
          },
          "normalize_volume": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Normalize Volume",
            "default": true
          }
        },
        "type": "object",
        "title": "AudioImportRequest",
        "description": "Audio import request model."
      },
      "AudioListResponse": {
        "properties": {
          "audio_files": {
            "items": {
              "$ref": "#/components/schemas/AudioResponse"
            },
            "type": "array",
            "title": "Audio Files"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "audio_files",
          "total"
        ],
        "title": "AudioListResponse",
        "description": "Audio list response model."
      },
      "AudioResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "duration": {
            "type": "number",
            "title": "Duration"
          },
          "format": {
            "type": "string",
            "title": "Format"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "sample_rate": {
            "type": "integer",
            "title": "Sample Rate"
          },
          "channels": {
            "type": "integer",
            "title": "Channels"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "waveform_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Waveform Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "name",
          "description",
          "duration",
          "format",
          "size",
          "sample_rate",
          "channels",
          "tags",
          "created_at",
          "url",
          "waveform_url"
        ],
        "title": "AudioResponse",
        "description": "Audio response model."
      },
      "Body_import_audio_audio_import_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          },
          "request": {
            "$ref": "#/components/schemas/AudioImportRequest",
            "default": {
              "auto_enhance": true,
              "normalize_volume": true
            }
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_import_audio_audio_import_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}